move humanize into common/utils
