+++ packages/common/src/index.js
@@ -14,7 +14,8 @@ import rpcClient from './services/rpcClient.js'
 import { ServicesProvider, useServices } from './contexts/ServicesContext.js'
 
 // utils
-import { getUrl } from './utils/urls.js'
+import { getUrl } from './utils/urls'
+import timeUtils from './utils/time'
 
 // components
 import Navbar from './components/Navbar'
@@ -30,7 +31,7 @@ import linesVector from './assets/images/lines.png'
 export {
   jwtService, rpcClient,
   ServicesProvider, useServices,
-  getUrl,
+  getUrl, timeUtils,
   Footer, FooterLink, Navbar, NavbarLink, Alert, Loading,
-  linesVector
+  linesVector,
 }
